+++ examples/filedownload_demo/workflows/test.workflow.js
@@ -1,0 +1,5 @@
+module.exports = {
+    workflow: [
+        'temp-usecase1'        
+    ]
+}